refactor(directive): extract register-protocol leaf to shrink v-register cycle
The v-register system was a 5-file import cycle: directive.ts (the hub)
plus directive-file, directive-lifecycle, use-register, and plugin, joined
by the register markers, the isRegisterValue / isTransforming guards, and
the assigner pipeline.

Extract the shared, leaf-eligible primitives into register-protocol.ts: the
V_REGISTER_MARKER / REGISTER_OWNER_MARKER markers and the isRegisterValue /
isTransforming guards (the leaf depends only on vue and the types module).
directive, directive-file, directive-lifecycle, and use-register now import
these from the leaf instead of from each other.

This removes use-register, plugin, and directive-lifecycle from the cycle:
the SCC shrinks from 5 files to 2 (directive <-> directive-file, still
coupled through the assigner pipeline <-> vRegisterFile, a cohesive pair
left visible rather than papered over). Fully splitting that last pair
would mean lifting the whole assigner pipeline out of directive.ts, a
larger change deferred for now.

Public API unchanged: src/index.ts still exports isRegisterValue (now from
the leaf), vRegister, and assignKey. The one test that imported
isRegisterValue from the old path follows the symbol; no assertions
changed. Full suite green (321 files / 3953 tests), matching the
pre-refactor baseline.
