+++ test/composables/coerce.test.ts
@@ -16,7 +16,8 @@ import { createApp, defineComponent, h, withDirectives, type App } from 'vue'
 import { z } from 'zod'
 import { useForm } from '../../src/zod'
 import type { UseFormConfigV4, UseFormReturnV4 } from '../../src/zod'
-import { vRegister, isRegisterValue, assignKey } from '../../src/runtime/core/directive'
+import { vRegister, assignKey } from '../../src/runtime/core/directive'
+import { isRegisterValue } from '../../src/runtime/core/register-protocol'
 import { createAttaform } from '../../src/runtime/core/plugin'
 import { defineCoercion, defaultCoercionRules } from '../../src/runtime/core/schema-coerce'
 import { awaitSettle, waitUntil } from '../utils/form-harness'
